requirements.txt
pandas
python-dotenv
numpy
ccxt
PySocks
requests[socks]
aiohttp-socks
openpyxl
fastapi
uvicorn
pydantic-settings
python-multipart
tqdm
